+++ core/docs/changelog.md
@@ -2,6 +2,9 @@
 ### Features
 * Added a test for the static file cache
 
+### Breaking Changes
+* `kurtosis.sh` no longer creates a suite execution volume, and the initializer container no longer accepts a param for it
+
 # 1.16.6
 ### Features
 * Moved the Kurtosis-internal testsuite from Kurtosis Libs into here, to break the circular dependency that used to exist between the two repos